+++ internal/backend/gs/gs.go
@@ -338,7 +338,7 @@ func (be *Backend) List(ctx context.Context, t backend.FileType, fn func(backend
 
 		fi := backend.FileInfo{
 			Name: path.Base(m),
-			Size: int64(attrs.Size),
+			Size: attrs.Size,
 		}
 
 		err = fn(fi)
