+++ internal/restorer/filerestorer.go
@@ -399,5 +399,5 @@ func (r *fileRestorer) reportBlobProgress(file *fileInfo, blobSize uint64) {
 	if file.state == nil {
 		action = restore.ActionFileRestored
 	}
-	r.progress.AddProgress(file.location, action, uint64(blobSize), uint64(file.size))
+	r.progress.AddProgress(file.location, action, blobSize, uint64(file.size))
 }
